requirements.txt
ultralytics>=8.0.0
SoccerNet>=3.0.0
opencv-python>=4.8.0
tqdm>=4.65.0
numpy>=1.24.0
pyyaml>=6.0

